Move funcitons to helper functions and fix imports
